requirements.txt
openai
pyyaml
rich
python-dotenv
